requirements.txt
flask==2.3.3
requests>=2.32.2
python-dotenv>=1.0.1
pydantic>=2.8.0
uvicorn>=0.30.1
fastapi>=0.116.1
cryptography>=3.4.8
ecdsa>=0.18.0
dateparser>=1.1.8
nltk>=3.8.1
openai>=1.0.0
anthropic>=0.7.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
uagents>=0.4.0
